Now able to create image of text on multiple lines
Added functionality to create an image of a block of text/sentence on multiple lines; the image is the images of all the lines and hence, the entire block of text/sentence. 

Also fixed a slight oversight where before, the variable lowest_baseline was being increased by a constant 20% but this only works for specific fonts. The variable lowest_baseline is now assigned the value of the word with the lowest baseline (= max y bottom value).

When creating the image of a block of text/sentence, the text can now be left-aligned, centered or right-aligned. The string of the block of text/sentence passed in as a parameter is also rearranged to accommodate for RTL nature of the Arabic language.