Optimized the code
Added a "cache" system so that all alphabets and all vowels won't be processed again from scratched for each word: images for each vowel were created everytime for a word even if these same images have been created for the previous word. This cache system primarily concerns the creation of the image of a block of text/sentence.

Modified function calculating width and height of text when drawn onto an image to also return the full bounding box; only left and top values were being returned previously. This eliminates the need to calculate values for the right and bottom of the bounding box by doing left + width, top + height, ...